requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
cloudpickle=0.5.3
cycler=0.10.0
cython=0.28.3
decorator=4.3.0
h5py=2.10.0
imageio=2.3.0
kiwisolver=1.1.0
matplotlib=3.1.1
networkx=2.1
ninja=1.9.0
numpy=1.16
pillow=6.1.0
pyqt=5.9.2
pyyaml=5.1.2
pywavelets=0.5.2
qt=5.9.7
scikit-image=0.15.0
scikit-learn=0.21.3
scipy=1.3.1
six=1.12.0
sqlite=3.29.0=h7b6447c_0
toolz=0.9.0
tqdm=4.31.1
wheel=0.33.6
yaml=0.1.7
